Move utility functions and constants out of cspad_tbx into serialtbx.detector.cspad and serialtbx.detector.xtc
